Can detect player fishing.
Caught items are not necessarily reliable however; theres no perfect way to tie an item to fishing. We check for item slot packets sent to the player immediately(250ms) following a fishing op-- but this could in theory get false positives/negatives.

No rewards or scoring for fishing are in place yet.