Added support for Scheduled scripts.
Scheduled scripts must go in the "scheduled" folder, where they will be automatically scanned for and loaded. Scheduled scripts should implement a GetSchedule() method, which returns a Scheduler object. 

Scheduler objectgs can be created via RunAt(), for specific times to run, or RunEvery(), to run a Script on a time interval. 

It is also possible to attach an optional condition, via the When() method.