CustomQuests: GatherItems trigger now accepts Action to run on gather
Action is of the form, MyAction(member as PartyMember, count as int). Only one ctor overload,  accepts the action( the most explicit overload ). 

Action execution takes place on the same thread that the quest runs on, since we dont want not to stall the thread that the ItemDrop event comes from.